move branch retrieval, createView and defaultBuildJob to JobHelper
