+++ jobs/jobs.groovy
@@ -1,4 +1,3 @@
-import groovy.json.JsonSlurper
 import helper.Globals
 import javaposse.jobdsl.dsl.jobs.FreeStyleJob
 import javaposse.jobdsl.dsl.views.ListView
@@ -183,8 +182,7 @@ void createBuildJobs(ListView view, String repo) {
  * Creates build jobs for feature-branches.
  */
 void createFeatureBranches(ListView view, String repo) {
-    def branchApi = new URL("https://api.github.com/repos/test-editor/$repo/branches")
-    def branches = new JsonSlurper().parse(branchApi.newReader())
+    def branches = getBranches(repo)
 
     branches.findAll { it.name.startsWith('feature/') }.each { branch ->
         def featureJobName = createJobName(repo, branch.name)
@@ -319,30 +317,4 @@ void createReleaseJobs4Fixtures(ListView view, String fixtureName, String repo) 
     addJob2View(view, releaseJobName)
 }
 
-/**
- * Creates list view with default columns.
- */
-ListView createView(String viewName, String text) {
-    return listView(viewName) {
-        description("${text}")
-        columns {
-            status()
-            weather()
-            name()
-            lastSuccess()
-            lastFailure()
-            lastDuration()
-            buildButton()
-        }
-    }
-}
 
-/**
- * Defines how a default build job should look like.
- */
-FreeStyleJob defaultBuildJob(String jobName, String repo, String branch, Closure closure) {
-    FreeStyleJob buildJob = job(jobName)
-    addDefaultConfiguration(buildJob, branch, closure)
-    addTEGitRepo(buildJob, repo, branch)
-    return buildJob
-}
